microservices/payment-service/requirements.txt
Flask==2.3.3
stripe==10.12.0
python-dotenv==1.0.1
pydantic==2.9.2